Move VisEditorContructor to visualize
