+++ src/plugins/visualizations/public/vis_types/types.ts
@@ -18,14 +18,12 @@
  */
 
 import { IconType } from '@elastic/eui';
-import { EventEmitter } from 'events';
 import React from 'react';
 import { Adapters } from 'src/plugins/inspector';
-import { IEditorController } from 'src/plugins/visualize/public';
+import { VisEditorConstructor } from 'src/plugins/visualize/public';
 import { ISchemas } from 'src/plugins/vis_default_editor/public';
 import { TriggerContextMapping } from '../../../ui_actions/public';
 import { Vis, VisToExpressionAst, VisualizationControllerConstructor } from '../types';
-import { VisualizeEmbeddableContract } from '..';
 
 export interface VisTypeOptions {
   showTimePicker: boolean;
@@ -34,13 +32,6 @@ export interface VisTypeOptions {
   showIndexSelection: boolean;
   hierarchicalData: boolean;
 }
-
-export type VisEditorConstructor = new (
-  element: HTMLElement,
-  vis: Vis,
-  eventEmitter: EventEmitter,
-  embeddableHandler: VisualizeEmbeddableContract
-) => IEditorController;
 
 /**
  * A visualization type representing one specific type of "classical"
