+++ src/plugins/visualize/public/index.ts
@@ -20,7 +20,11 @@
 import { PluginInitializerContext } from 'kibana/public';
 import { VisualizePlugin } from './plugin';
 
-export { EditorRenderProps, IEditorController } from './application/types';
+export type {
+  EditorRenderProps,
+  IEditorController,
+  VisEditorConstructor,
+} from './application/types';
 export { VisualizeConstants } from './application/visualize_constants';
 
 export const plugin = (context: PluginInitializerContext) => {
